Align submission statuses in code and tests
The names of submission statuses in tests have got slightly out of sync
with those used in code.

This commit renames the 'in_progess' status to 'processing' in the
Submission factory and updates tests appropriately.

It also adds the 'failed' status so that the factory fully reflects the
possible statuses of a submission.