Tech-Debt: Remove use case endpoints from swagger docs
These were the first method of calling the interface and have been
replaced by the submission/create and result endpoints.  They were
only semi-functional now due to changes in the process and should
have been removed a while ago.  As we are close to allowing other
users access to the documentation, I am removing them now to prevent
confusion and errors if they were called accidentally