Amend smoke test start and end dates
The HMRC sandbox no longer returns data for
the period specified but does for the period
as it was created with data from
```
spec/fixtures/test_data/paye.json
```
Posibly it was changes to the sandbox that rely on `paymentDate`
have meant we now have to query a date range that covers these
payment dates.

From developer hub, found by Colin.
> The paymentDate value provided in the POST body will be used to determine if the PAYE data sits within the date range provided in the query parameters, if they do not then the PAYE data will not be returned.