Remove smoke test values from old secrets
In dicussion with team it was determined that these
were only made secrets because initially realistic data
had been used to test the API but that since then we have
generated and used fake data. The current data being supplied
to the smoke tests is entirely fake and therefore does not need
to be treated as a secret.